requirements.txt
albumentations==1.1.0
imgaug==0.4.0
matplotlib==3.5.1
numpy
opencv-python==4.5.1.48
pandas==1.1.5
Pillow==8.4.0
tensorboard==2.8.0
torch==1.10.1
torchaudio==0.10.1
torchvision==0.11.2
tqdm==4.62.3